+++ addon/options.js
@@ -1,41 +1,41 @@
 /******/ (function(modules) { // webpackBootstrap
 /******/ 	// The module cache
 /******/ 	var installedModules = {};
-
+/******/
 /******/ 	// The require function
 /******/ 	function __webpack_require__(moduleId) {
-
+/******/
 /******/ 		// Check if module is in cache
-/******/ 		if(installedModules[moduleId])
+/******/ 		if(installedModules[moduleId]) {
 /******/ 			return installedModules[moduleId].exports;
-
+/******/ 		}
 /******/ 		// Create a new module (and put it into the cache)
 /******/ 		var module = installedModules[moduleId] = {
 /******/ 			i: moduleId,
 /******/ 			l: false,
 /******/ 			exports: {}
 /******/ 		};
-
+/******/
 /******/ 		// Execute the module function
 /******/ 		modules[moduleId].call(module.exports, module, module.exports, __webpack_require__);
-
+/******/
 /******/ 		// Flag the module as loaded
 /******/ 		module.l = true;
-
+/******/
 /******/ 		// Return the exports of the module
 /******/ 		return module.exports;
 /******/ 	}
-
-
+/******/
+/******/
 /******/ 	// expose the modules object (__webpack_modules__)
 /******/ 	__webpack_require__.m = modules;
-
+/******/
 /******/ 	// expose the module cache
 /******/ 	__webpack_require__.c = installedModules;
-
+/******/
 /******/ 	// identity function for calling harmony imports with the correct context
 /******/ 	__webpack_require__.i = function(value) { return value; };
-
+/******/
 /******/ 	// define getter function for harmony exports
 /******/ 	__webpack_require__.d = function(exports, name, getter) {
 /******/ 		if(!__webpack_require__.o(exports, name)) {
@@ -46,7 +46,7 @@
 /******/ 			});
 /******/ 		}
 /******/ 	};
-
+/******/
 /******/ 	// getDefaultExport function for compatibility with non-harmony modules
 /******/ 	__webpack_require__.n = function(module) {
 /******/ 		var getter = module && module.__esModule ?
@@ -55,239 +55,74 @@
 /******/ 		__webpack_require__.d(getter, 'a', getter);
 /******/ 		return getter;
 /******/ 	};
-
+/******/
 /******/ 	// Object.prototype.hasOwnProperty.call
 /******/ 	__webpack_require__.o = function(object, property) { return Object.prototype.hasOwnProperty.call(object, property); };
-
+/******/
 /******/ 	// __webpack_public_path__
 /******/ 	__webpack_require__.p = "/";
-
+/******/
 /******/ 	// Load entry module and return exports
-/******/ 	return __webpack_require__(__webpack_require__.s = 16);
+/******/ 	return __webpack_require__(__webpack_require__.s = 17);
 /******/ })
 /************************************************************************/
 /******/ ([
 /* 0 */
 /***/ (function(module, exports, __webpack_require__) {
 
-"use strict";
-
-
-var env = __webpack_require__(1);
-
-function handleResponse(r) {
-    console.info(r);
-}
-
-function toTitleCase(word) {
-    return word.charAt(0).toUpperCase() + word.substr(1).toLowerCase();
-}
-
-function getLocalString(key) {
-    return browser.i18n.getMessage(key);
-}
-
-function log() {
-    if (env.debugMode) {
-        for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {
-            args[_key] = arguments[_key];
-        }
-
-        console.log(args.reduce(function (acc, val) {
-            return acc + val;
-        }, ''));
-    }
-}
-
-function formatTime(minutes) {
-    if (!Number.isInteger(minutes)) {
-        console.error('input time [' + minutes + '] is not integer');
-        return 'ERROR';
-    }
-    var formatted = [];
-    var hour = 0;
-    if (minutes >= 60) {
-        do {
-            minutes -= 60;
-            hour += 1;
-        } while (minutes >= 60);
-    }
-    formatted.push(hour, minutes);
-    return formatted.map(function (el) {
-        return padTime(el);
-    }).join(':');
-}
-
-function padTime(val) {
-    return val < 10 ? '0' + val : val;
-}
-
-module.exports = {
-    handleResponse: handleResponse,
-    toTitleCase: toTitleCase,
-    getLocalString: getLocalString,
-    log: log,
-    formatTime: formatTime
-};
+const env=__webpack_require__(2);function handleResponse(a){console.info(a)}function toTitleCase(a){return a.charAt(0).toUpperCase()+a.substr(1).toLowerCase()}function getLocalString(a){return browser.i18n.getMessage(a)}function log(...a){env.debugMode&&console.log(a.reduce((b,c)=>b+c,''))}function formatTime(a){if(!Number.isInteger(a))return console.error(`input time [${a}] is not integer`),'ERROR';let b=[],c=0;if(60<=a)do a-=60,c+=1;while(60<=a);return b.push(c,a),b.map((d)=>padTime(d)).join(':')}function padTime(a){return 10>a?`0${a}`:a}module.exports={handleResponse,toTitleCase,getLocalString,log,formatTime};
 
 /***/ }),
 /* 1 */
-/***/ (function(module, exports, __webpack_require__) {
+/***/ (function(module, exports) {
 
-"use strict";
-
-
-module.exports = {
-    debugMode: true
-};
+module.exports={isReading:!0,passedMinutes:0,breakTimeAmount:10,readingTimeAmount:50,idleDetectionInterval:1200,title:browser.i18n.getMessage("notificationTitle"),message:browser.i18n.getMessage("notificationMessage")};
 
 /***/ }),
-/* 2 */,
-/* 3 */
-/***/ (function(module, exports, __webpack_require__) {
+/* 2 */
+/***/ (function(module, exports) {
 
-"use strict";
-
-
-var nodes = {
-    pool: {},
-    getDOM: function getDOM(key) {
-        if (nodes.pool.hasOwnProperty(key)) {
-            return nodes.pool[key];
-        } else {
-            return nodes.pool[key] = document.querySelector('#' + key);
-        }
-    }
-};
-
-module.exports = nodes;
+module.exports={debugMode:!0};
 
 /***/ }),
+/* 3 */,
 /* 4 */
-/***/ (function(module, exports, __webpack_require__) {
+/***/ (function(module, exports) {
 
-"use strict";
-
-
-module.exports = {
-    isReading: true,
-    passedMinutes: 0,
-    breakTimeAmount: 10,
-    readingTimeAmount: 50,
-    idleDetectionInterval: 1200
-};
+var nodes={pool:{},getDOM(a){return nodes.pool.hasOwnProperty(a)?nodes.pool[a]:nodes.pool[a]=document.querySelector('#'+a)}};module.exports=nodes;
 
 /***/ }),
 /* 5 */,
 /* 6 */,
 /* 7 */,
-/* 8 */
+/* 8 */,
+/* 9 */
 /***/ (function(module, exports, __webpack_require__) {
 
-"use strict";
-
-
-var _require = __webpack_require__(0),
-    toTitleCase = _require.toTitleCase,
-    getLocalString = _require.getLocalString;
-
-var defaultValues = __webpack_require__(4);
-var nodes = __webpack_require__(3);
-
-var reflect = {
-    readingTime_label: 'optionsWorkTimeLabel',
-    breakTime_label: 'optionsBreakTimeLabel',
-    apply: 'optionsApplyButton'
-};
-var storageKeys = ['breakTimeAmount', 'readingTimeAmount'];
-
-var page = {
-    render: function render() {
-        browser.storage.local.get(storageKeys).then(page.inputs.set, function (err) {
-            console.error(err);
-        });
-
-        for (var key in reflect) {
-            nodes.getDOM(key).innerText = getLocalString(reflect[key]);
-        }
-    },
-
-    inputs: {
-        set: function set(result) {
-            storageKeys.forEach(function (key) {
-                nodes.getDOM(key).value = result[key] || defaultValues[key];
-            });
-        },
-        get: function get() {
-            var set = {
-                last_modified: new Date().toLocaleTimeString()
-            };
-            storageKeys.forEach(function (key) {
-                set[key] = parseInt(nodes.getDOM(key).value);
-            });
-            return set;
-        }
-    }
-};
-
-module.exports = page;
+const{toTitleCase,getLocalString}=__webpack_require__(0),defaultValues=__webpack_require__(1);var nodes=__webpack_require__(4);const reflect={readingTime_label:'optionsWorkTimeLabel',breakTime_label:'optionsBreakTimeLabel',title_label:'optionsNotificationTitleLabel',message_label:'optionsNotificationMessageLabel',apply:'optionsApplyButton',reset:'optionsResetButton'},storageKeys=['breakTimeAmount','readingTimeAmount','title','message'];var page={render(){page.labels.set(),page.inputs.set()},labels:{set(){for(var a in reflect)nodes.getDOM(a).innerText=getLocalString(reflect[a])}},inputs:{set(){browser.storage.local.get(storageKeys).then((a)=>{storageKeys.forEach((b)=>{nodes.getDOM(b).value=a[b]||defaultValues[b]})}).catch((a)=>{console.error(a)})},get(){let a={last_modified:new Date().toLocaleTimeString()};return storageKeys.forEach((b)=>{let c=nodes.getDOM(b),d=c.attributes;a[b]=d.type&&'number'===d.type.value?c.valueAsNumber:c.value}),a},dom(){return Array.from(storageKeys.map((a)=>nodes.getDOM(a)))}}};module.exports=page;
 
 /***/ }),
-/* 9 */,
-/* 10 */
+/* 10 */,
+/* 11 */
 /***/ (function(module, exports, __webpack_require__) {
 
-"use strict";
-
-
-var _require = __webpack_require__(0),
-    getLocalString = _require.getLocalString;
-
-var page = __webpack_require__(8);
-var nodes = __webpack_require__(3);
-
-var options = {
-    save: function save(e) {
-        e.preventDefault();
-        var data = page.inputs.get();
-        var time = data.last_modified;
-
-        browser.storage.local.set(data).then(function () {
-            nodes.getDOM('apply_msg').innerText = '[' + time + '] ' + getLocalString('optionsApplySuccessMessage');
-            nodes.getDOM('error_msg').innerText = '';
-        }).catch(function (err) {
-            nodes.getDOM('apply_msg').innerText = '';
-            nodes.getDOM('error_msg').innerText = '[' + time + '] ' + err.message;
-        });
-    }
-};
-
-module.exports = options;
+const{getLocalString}=__webpack_require__(0);var page=__webpack_require__(9),nodes=__webpack_require__(4),defaultValues=__webpack_require__(1),options={save(a){a.preventDefault();let b=page.inputs.get(),c=page.inputs.dom(),d=`[${b.last_modified}] `;return options.checkInputErrorExists(c)?void options.msg('',!1):void options.apply(d,b)},reset(a){a.preventDefault();let b=({title,message,breakTimeAmount,readingTimeAmount}=defaultValues),c=b.last_modified=new Date().toLocaleTimeString();options.apply(c,b,function(){page.inputs.set()})},checkInputErrorExists(a){return a.map((b)=>{return b.attributes.type&&'number'===b.attributes.type.value&&(b.validity.rangeOverflow||b.validity.rangeUnderflow)}).filter((b)=>b).length},apply(a,b,c=void 0){browser.storage.local.set(b).then(()=>{options.msg(a+getLocalString('optionsApplySuccessMessage'),!0),c&&c()}).catch((d)=>{options.msg(a+d.message,!1)})},msg(a,b){let c=nodes.getDOM('msg');c.innerText=a,c.classList.toggle('success',b),c.classList.toggle('warning',!b)}};module.exports=options;
 
 /***/ }),
-/* 11 */,
-/* 12 */
+/* 12 */,
+/* 13 */
 /***/ (function(module, exports) {
 
 // removed by extract-text-webpack-plugin
 
 /***/ }),
-/* 13 */,
 /* 14 */,
 /* 15 */,
-/* 16 */
+/* 16 */,
+/* 17 */
 /***/ (function(module, exports, __webpack_require__) {
 
-"use strict";
-
-
-__webpack_require__(12);
-var options = __webpack_require__(10);
-var page = __webpack_require__(8);
-var nodes = __webpack_require__(3);
-
-document.addEventListener("DOMContentLoaded", page.render);
-nodes.getDOM("apply").addEventListener("click", options.save);
+__webpack_require__(13);var options=__webpack_require__(11),page=__webpack_require__(9),nodes=__webpack_require__(4);document.addEventListener('DOMContentLoaded',page.render),nodes.getDOM('apply').addEventListener('click',options.save),nodes.getDOM('reset').addEventListener('click',options.reset);
 
 /***/ })
 /******/ ]);