+++ addon/popup.js
@@ -1,41 +1,41 @@
 /******/ (function(modules) { // webpackBootstrap
 /******/ 	// The module cache
 /******/ 	var installedModules = {};
-
+/******/
 /******/ 	// The require function
 /******/ 	function __webpack_require__(moduleId) {
-
+/******/
 /******/ 		// Check if module is in cache
-/******/ 		if(installedModules[moduleId])
+/******/ 		if(installedModules[moduleId]) {
 /******/ 			return installedModules[moduleId].exports;
-
+/******/ 		}
 /******/ 		// Create a new module (and put it into the cache)
 /******/ 		var module = installedModules[moduleId] = {
 /******/ 			i: moduleId,
 /******/ 			l: false,
 /******/ 			exports: {}
 /******/ 		};
-
+/******/
 /******/ 		// Execute the module function
 /******/ 		modules[moduleId].call(module.exports, module, module.exports, __webpack_require__);
-
+/******/
 /******/ 		// Flag the module as loaded
 /******/ 		module.l = true;
-
+/******/
 /******/ 		// Return the exports of the module
 /******/ 		return module.exports;
 /******/ 	}
-
-
+/******/
+/******/
 /******/ 	// expose the modules object (__webpack_modules__)
 /******/ 	__webpack_require__.m = modules;
-
+/******/
 /******/ 	// expose the module cache
 /******/ 	__webpack_require__.c = installedModules;
-
+/******/
 /******/ 	// identity function for calling harmony imports with the correct context
 /******/ 	__webpack_require__.i = function(value) { return value; };
-
+/******/
 /******/ 	// define getter function for harmony exports
 /******/ 	__webpack_require__.d = function(exports, name, getter) {
 /******/ 		if(!__webpack_require__.o(exports, name)) {
@@ -46,7 +46,7 @@
 /******/ 			});
 /******/ 		}
 /******/ 	};
-
+/******/
 /******/ 	// getDefaultExport function for compatibility with non-harmony modules
 /******/ 	__webpack_require__.n = function(module) {
 /******/ 		var getter = module && module.__esModule ?
@@ -55,15 +55,15 @@
 /******/ 		__webpack_require__.d(getter, 'a', getter);
 /******/ 		return getter;
 /******/ 	};
-
+/******/
 /******/ 	// Object.prototype.hasOwnProperty.call
 /******/ 	__webpack_require__.o = function(object, property) { return Object.prototype.hasOwnProperty.call(object, property); };
-
+/******/
 /******/ 	// __webpack_public_path__
 /******/ 	__webpack_require__.p = "/";
-
+/******/
 /******/ 	// Load entry module and return exports
-/******/ 	return __webpack_require__(__webpack_require__.s = 17);
+/******/ 	return __webpack_require__(__webpack_require__.s = 18);
 /******/ })
 /************************************************************************/
 /******/ ({
@@ -71,150 +71,35 @@
 /***/ 0:
 /***/ (function(module, exports, __webpack_require__) {
 
-"use strict";
-
-
-var env = __webpack_require__(1);
-
-function handleResponse(r) {
-    console.info(r);
-}
-
-function toTitleCase(word) {
-    return word.charAt(0).toUpperCase() + word.substr(1).toLowerCase();
-}
-
-function getLocalString(key) {
-    return browser.i18n.getMessage(key);
-}
-
-function log() {
-    if (env.debugMode) {
-        for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {
-            args[_key] = arguments[_key];
-        }
-
-        console.log(args.reduce(function (acc, val) {
-            return acc + val;
-        }, ''));
-    }
-}
-
-function formatTime(minutes) {
-    if (!Number.isInteger(minutes)) {
-        console.error('input time [' + minutes + '] is not integer');
-        return 'ERROR';
-    }
-    var formatted = [];
-    var hour = 0;
-    if (minutes >= 60) {
-        do {
-            minutes -= 60;
-            hour += 1;
-        } while (minutes >= 60);
-    }
-    formatted.push(hour, minutes);
-    return formatted.map(function (el) {
-        return padTime(el);
-    }).join(':');
-}
-
-function padTime(val) {
-    return val < 10 ? '0' + val : val;
-}
-
-module.exports = {
-    handleResponse: handleResponse,
-    toTitleCase: toTitleCase,
-    getLocalString: getLocalString,
-    log: log,
-    formatTime: formatTime
-};
+const env=__webpack_require__(2);function handleResponse(a){console.info(a)}function toTitleCase(a){return a.charAt(0).toUpperCase()+a.substr(1).toLowerCase()}function getLocalString(a){return browser.i18n.getMessage(a)}function log(...a){env.debugMode&&console.log(a.reduce((b,c)=>b+c,''))}function formatTime(a){if(!Number.isInteger(a))return console.error(`input time [${a}] is not integer`),'ERROR';let b=[],c=0;if(60<=a)do a-=60,c+=1;while(60<=a);return b.push(c,a),b.map((d)=>padTime(d)).join(':')}function padTime(a){return 10>a?`0${a}`:a}module.exports={handleResponse,toTitleCase,getLocalString,log,formatTime};
 
 /***/ }),
 
-/***/ 1:
+/***/ 12:
 /***/ (function(module, exports, __webpack_require__) {
 
-"use strict";
-
-
-module.exports = {
-    debugMode: true
-};
+const{formatTime}=__webpack_require__(0);var clock={dom:null,request(){browser.runtime.sendMessage({type:'requestTime'}).then(clock.update).catch((a)=>{console.error(a)})},reset(){browser.runtime.sendMessage({type:'resetCounter'}).then(clock.update).catch((a)=>{console.error(a)})},update(a){if('object'==typeof a){clock.dom||(clock.dom=document.querySelector('#monitor'));let b=a.limit-a.time;clock.dom.innerText=formatTime(b),clock.dom.classList.toggle('warning',!a.reading)}return!0}};module.exports=clock;
 
 /***/ }),
 
-/***/ 11:
-/***/ (function(module, exports, __webpack_require__) {
-
-"use strict";
-
-
-var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol" ? function (obj) { return typeof obj; } : function (obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; };
-
-var _require = __webpack_require__(0),
-    formatTime = _require.formatTime;
-
-var clock = {
-    dom: null,
-    request: function request() {
-        browser.runtime.sendMessage({ type: 'requestTime' }).then(clock.update).catch(function (err) {
-            console.error(err);
-        });
-    },
-    reset: function reset() {
-        browser.runtime.sendMessage({ type: 'resetCounter' }).then(clock.update).catch(function (err) {
-            console.error(err);
-        });
-    },
-    update: function update(msg) {
-        if ((typeof msg === 'undefined' ? 'undefined' : _typeof(msg)) === 'object') {
-            if (!clock.dom) {
-                clock.dom = document.querySelector("#monitor");
-            }
-            clock.dom.innerText = formatTime(msg.time);
-            clock.dom.classList.toggle('warning', !msg.reading);
-        }
-        return true;
-    }
-};
-
-module.exports = clock;
-
-/***/ }),
-
-/***/ 13:
+/***/ 14:
 /***/ (function(module, exports) {
 
 // removed by extract-text-webpack-plugin
 
 /***/ }),
 
-/***/ 17:
+/***/ 18:
 /***/ (function(module, exports, __webpack_require__) {
 
-"use strict";
+__webpack_require__(14);var clock=__webpack_require__(12);document.querySelector("#refresh_button").addEventListener("click",clock.reset),document.querySelector("#options_button").addEventListener("click",(a)=>{a.preventDefault(),browser.runtime.openOptionsPage()}),window.addEventListener("load",function(){browser.runtime.onMessage.addListener(clock.update),clock.request()}),window.addEventListener("unload",function(){browser.runtime.onMessage.removeListener(clock.update)});
 
+/***/ }),
 
-__webpack_require__(13);
-var clock = __webpack_require__(11);
+/***/ 2:
+/***/ (function(module, exports) {
 
-document.querySelector('#refresh_button').addEventListener('click', clock.reset);
-
-document.querySelector('#options_button').addEventListener('click', function (e) {
-    e.preventDefault();
-    browser.runtime.openOptionsPage();
-});
-
-window.addEventListener("load", function (event) {
-    browser.runtime.onMessage.addListener(clock.update);
-    clock.request();
-});
-
-window.addEventListener("unload", function (event) {
-    browser.runtime.onMessage.removeListener(clock.update);
-});
+module.exports={debugMode:!0};
 
 /***/ })