+++ src/background_scripts/main.js
@@ -4,13 +4,14 @@ var idle = require("./idle")
 var clock = require("./clock")
 var counter = require("./counter")
 var storage = require("./storage")
+var timePacket = require("./timePacket")
 
 /**
  *  callbacks
  */
 function resetUI(iconIsGreen) {
     clock.reset()
-    ui.icon.switch(iconIsGreen)
+    //ui.icon.switch(iconIsGreen)
     ui.clock.switch(iconIsGreen)
     ui.clock.sync()
 }
@@ -44,23 +45,16 @@ browser.alarms.onAlarm.addListener(alarm => {
 })
 
 // start dispatch request
-
-function timeResponse() {
-    return {
-        time: storage.store.passedMinutes,
-        reading: storage.store.isReading
-    }
-}
 browser.runtime.onMessage.addListener((request, sender) => {
     switch (request.type) {
 
         case 'requestTime':
-            return Promise.resolve(timeResponse())
+            return Promise.resolve(timePacket())
 
         case 'resetCounter':
             counter.restart()
             resetUI(true)
-            return Promise.resolve(timeResponse())
+            return Promise.resolve(timePacket())
     }
 })