+++ src/background_scripts/timePacket.js
@@ -1,0 +1,11 @@
+var storage = require("./storage")
+
+var timePacket = () => {
+    return {
+        time: storage.store.passedMinutes,
+        limit: storage.store.isReading ? storage.store.readingTimeAmount : storage.store.breakTimeAmount,
+        reading: storage.store.isReading
+    }
+}
+
+module.exports = timePacket