+++ src/options/main.js
@@ -5,3 +5,4 @@ var nodes = require('./nodes')
 
 document.addEventListener("DOMContentLoaded", page.render);
 nodes.getDOM("apply").addEventListener("click", options.save);
+nodes.getDOM("reset").addEventListener("click", options.reset);