+++ src/options/page.js
@@ -3,43 +3,62 @@ const defaultValues = require('../configs/defaults')
 var nodes = require('./nodes')
 
 const reflect = {
-    /**
-     * [dom key]: [locale string key]
-     */
+    // [dom key]: [locale string key]
     readingTime_label: 'optionsWorkTimeLabel',
     breakTime_label: 'optionsBreakTimeLabel',
-    apply: 'optionsApplyButton'
+    title_label: 'optionsNotificationTitleLabel',
+    message_label: 'optionsNotificationMessageLabel',
+    apply: 'optionsApplyButton',
+    reset: 'optionsResetButton'
 }
 const storageKeys = [
     'breakTimeAmount',
-    'readingTimeAmount'
+    'readingTimeAmount',
+    'title',
+    'message'
 ]
 
 var page = {
     render() {
-        // input
-        browser.storage.local.get(storageKeys)
-            .then(page.inputs.set, err => { console.error(err) })
-
-        // label
-        for (var key in reflect) {
-            nodes.getDOM(key).innerText = getLocalString(reflect[key])
+        page.labels.set()
+        page.inputs.set()
+    },
+    labels: {
+        set() {
+            for (var key in reflect) {
+                nodes.getDOM(key).innerText = getLocalString(reflect[key])
+            }
         }
     },
     inputs: {
-        set(result) {
-            storageKeys.forEach(key => {
-                nodes.getDOM(key).value = result[key] || defaultValues[key]
-            })
+        set() {
+            browser.storage.local.get(storageKeys)
+                .then(result => {
+                    storageKeys.forEach(key => {
+                        nodes.getDOM(key).value = result[key] || defaultValues[key]
+                    })
+                })
+                .catch(err => {
+                    console.error(err)
+                })
         },
         get() {
             let set = {
                 last_modified: (new Date).toLocaleTimeString()
             }
             storageKeys.forEach(key => {
-                set[key] = parseInt(nodes.getDOM(key).value)
+                let node = nodes.getDOM(key)
+                let attrs = node.attributes
+                if (attrs.type && attrs.type.value === 'number') {
+                    set[key] = node.valueAsNumber
+                } else {
+                    set[key] = node.value
+                }
             })
             return set
+        },
+        dom() {
+            return Array.from(storageKeys.map(key => nodes.getDOM(key)))
         }
     }
 }